test(ci): rm gfortran 13, 14
Although the band_distribution unit tests pass locally with
gfortran 13 and 14, some issue currently causes the CI tests
to fail with these compiler versions.  This commit removes the
problematic versions and retains the CI testing with gfortran 15,
flang 19-21, and ifx 2025.2.0 and 2025.2.1. We can reintroduce
CI testing with older versions of gfortrana later if so desired.